LPS-77350 configure gradle/maven tasks to support service.xml in resources folder
In order to get the service.xml files in modules to be read by our
bnd plugin, we need to make it easier to get the service.xml files into
the output bundle jar.  The easiest way is to simply put the service.xml
file into the src/main/resources/META-INF folder.  That can be the new
default location.

If it is not there we can fallback to the previous location for service.xml
but the developer will have to do additional work to get the service.xml
added to the bundle (-includeresource) and then it can be processed by our
ServiceAnalyzerPlugin